LayoutTests/fast/css-grid-layout/named-grid-line-get-set-expected.txt

Test that setting and getting grid-definition-columns and grid-definition-rows works as expected

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Test getting -webkit-grid-definition-columns and -webkit-grid-definition-rows set through CSS
PASS getComputedStyle(gridElement, '').getPropertyValue('-webkit-grid-definition-columns') is "first 10px"
PASS getComputedStyle(gridElement, '').getPropertyValue('-webkit-grid-definition-rows') is "first 15px"
PASS getComputedStyle(gridElement, '').getPropertyValue('-webkit-grid-definition-columns') is "53% last"
PASS getComputedStyle(gridElement, '').getPropertyValue('-webkit-grid-definition-rows') is "27% last"
PASS getComputedStyle(gridElement, '').getPropertyValue('-webkit-grid-definition-columns') is "first auto"
PASS getComputedStyle(gridElement, '').getPropertyValue('-webkit-grid-definition-rows') is "auto last"
PASS getComputedStyle(gridElement, '').getPropertyValue('-webkit-grid-definition-columns') is "first minmax(10%, 15px)"
PASS getComputedStyle(gridElement, '').getPropertyValue('-webkit-grid-definition-rows') is "minmax(20px, 50%) last"
PASS getComputedStyle(gridElement, '').getPropertyValue('-webkit-grid-definition-columns') is "nav first 10px last"
PASS getComputedStyle(gridElement, '').getPropertyValue('-webkit-grid-definition-rows') is "nav first 15px last"
PASS getComputedStyle(gridElement, '').getPropertyValue('-webkit-grid-definition-columns') is "nav first 10% nav 15% last"
PASS getComputedStyle(gridElement, '').getPropertyValue('-webkit-grid-definition-rows') is "first nav2 25% nav2 75% last"

Test getting and setting -webkit-grid-definition-columns and -webkit-grid-definition-rows through JS
PASS getComputedStyle(gridElement, '').getPropertyValue('-webkit-grid-definition-columns') is "first 18px"
PASS getComputedStyle(gridElement, '').getPropertyValue('-webkit-grid-definition-rows') is "66px last"
PASS getComputedStyle(gridElement, '').getPropertyValue('-webkit-grid-definition-columns') is "first 55%"
PASS getComputedStyle(gridElement, '').getPropertyValue('-webkit-grid-definition-rows') is "40% last"
PASS getComputedStyle(gridElement, '').getPropertyValue('-webkit-grid-definition-columns') is "first auto"
PASS getComputedStyle(gridElement, '').getPropertyValue('-webkit-grid-definition-rows') is "auto last"
PASS getComputedStyle(gridElement, '').getPropertyValue('-webkit-grid-definition-columns') is "first -webkit-min-content"
PASS getComputedStyle(gridElement, '').getPropertyValue('-webkit-grid-definition-rows') is "-webkit-min-content last"
PASS getComputedStyle(gridElement, '').getPropertyValue('-webkit-grid-definition-columns') is "first -webkit-max-content"
PASS getComputedStyle(gridElement, '').getPropertyValue('-webkit-grid-definition-rows') is "-webkit-max-content last"
PASS getComputedStyle(gridElement, '').getPropertyValue('-webkit-grid-definition-columns') is "first minmax(55%, 45px)"
PASS getComputedStyle(gridElement, '').getPropertyValue('-webkit-grid-definition-rows') is "minmax(30px, 40%) last"
PASS getComputedStyle(gridElement, '').getPropertyValue('-webkit-grid-definition-columns') is "first minmax(220px, -webkit-max-content)"
PASS getComputedStyle(gridElement, '').getPropertyValue('-webkit-grid-definition-rows') is "minmax(-webkit-max-content, 50px) last"
PASS getComputedStyle(gridElement, '').getPropertyValue('-webkit-grid-definition-columns') is "first minmax(220px, -webkit-max-content)"
PASS getComputedStyle(gridElement, '').getPropertyValue('-webkit-grid-definition-rows') is "minmax(-webkit-max-content, 50px) last"
PASS getComputedStyle(gridElement, '').getPropertyValue('-webkit-grid-definition-columns') is "first minmax(-webkit-min-content, -webkit-max-content)"
PASS getComputedStyle(gridElement, '').getPropertyValue('-webkit-grid-definition-rows') is "minmax(-webkit-max-content, -webkit-min-content) last"
PASS getComputedStyle(gridElement, '').getPropertyValue('-webkit-grid-definition-columns') is "nav first minmax(-webkit-min-content, -webkit-max-content) last"
PASS getComputedStyle(gridElement, '').getPropertyValue('-webkit-grid-definition-rows') is "nav first minmax(-webkit-max-content, -webkit-min-content) last"
PASS getComputedStyle(gridElement, '').getPropertyValue('-webkit-grid-definition-columns') is "nav first minmax(-webkit-min-content, -webkit-max-content) nav auto last"
PASS getComputedStyle(gridElement, '').getPropertyValue('-webkit-grid-definition-rows') is "first nav2 minmax(-webkit-max-content, -webkit-min-content) nav2 minmax(10px, 15px) last"
PASS getComputedStyle(gridElement, '').getPropertyValue('-webkit-grid-definition-columns') is "foo bar auto foo auto bar"
PASS getComputedStyle(gridElement, '').getPropertyValue('-webkit-grid-definition-rows') is "foo bar auto foo auto bar"

Test getting and setting invalid -webkit-grid-definition-columns and -webkit-grid-definition-rows through JS
PASS getComputedStyle(gridElement, '').getPropertyValue('-webkit-grid-definition-columns') is "none"
PASS getComputedStyle(gridElement, '').getPropertyValue('-webkit-grid-definition-rows') is "none"
PASS successfullyParsed is true

TEST COMPLETE
